+++ crates/runtime/src/search/full_text/index.rs
@@ -16,17 +16,7 @@ limitations under the License.
 
 use std::{any::Any, sync::Arc};
 
-use arrow::{
-    array::{
-        Array, ArrayRef, BinaryArray, BooleanArray, Date32Array, Date64Array, Float16Array,
-        Float32Array, Float64Array, Int8Array, Int16Array, Int32Array, Int64Array,
-        LargeBinaryArray, LargeStringArray, RecordBatch, StringArray, UInt8Array, UInt16Array,
-        UInt32Array, UInt64Array,
-    },
-    datatypes::DataType,
-    error::ArrowError,
-};
-use tantivy::{Term, schema::Field};
+use arrow::{array::RecordBatch, datatypes::DataType};
 
 use async_trait::async_trait;
 use datafusion::datasource::TableProvider;
@@ -43,7 +33,10 @@ use std::collections::HashSet;
 use tantivy::schema::DocParsingError;
 use tantivy::{TantivyDocument, TantivyError};
 
-use crate::search::util::get_primary_keys;
+use crate::search::{
+    full_text::util::{array_to_terms, with_json_subset_column},
+    util::get_primary_keys,
+};
 use crate::{datafusion::query::write_to_json_string, search::full_text::Error};
 use crate::{
     object_store_registry::SpiceObjectStoreRegistry,
@@ -54,7 +47,7 @@ use crate::{
 
 /// The minimum number of bytes to support writing to in-memory [`tantivy::Index`].
 pub static MINIMUM_MEMORY_BUDGET_FOR_MEMORY_INDEX: usize = 15_000_000;
-pub static INDEX_UNIQUE_FIELD_NAME: &'static str = "__spice.unique_field";
+pub static INDEX_UNIQUE_FIELD_NAME: &str = "__spice.unique_field";
 
 #[derive(Clone, Debug)]
 pub struct FullTextDatabaseIndex {
@@ -104,9 +97,9 @@ impl FullTextDatabaseIndex {
         };
 
         // INDEX_UNIQUE_FIELD_NAME is a reserved field name.
-        if pks.contains(INDEX_UNIQUE_FIELD_NAME) {
+        if pks.contains(&INDEX_UNIQUE_FIELD_NAME.to_string()) {
             return Err(Error::PrimaryKeyInvalidName {
-                column: INDEX_UNIQUE_FIELD_NAME,
+                column: INDEX_UNIQUE_FIELD_NAME.to_string(),
             });
         }
 
@@ -174,8 +167,18 @@ impl FullTextDatabaseIndex {
             index_writer.delete_term(t);
         }
 
-        // TODO: need to add INDEX_UNIQUE_FIELD_NAME when needed.
-        let doc_json = write_to_json_string(rb).context(InvalidIndexingSnafu {
+        // Construct column for `INDEX_UNIQUE_FIELD_NAME` if needed.
+        let rb = if self.primary_key.len() > 1 {
+            rb.iter()
+                .map(|r| with_json_subset_column(r, &self.primary_key, INDEX_UNIQUE_FIELD_NAME))
+                .collect::<Result<Vec<RecordBatch>, _>>()
+                .context(InvalidIndexingSnafu {
+                    context: "An error occured creating the a unique column for the full text search index".to_string(),
+                })?
+        } else {
+            rb.to_vec()
+        };
+        let doc_json = write_to_json_string(&rb).context(InvalidIndexingSnafu {
             context: "Failed to write data to intermediate JSON string for indexing".to_string(),
         })?;
         let docs = parse_json_array(&self.index.schema(), doc_json.as_str())
@@ -337,197 +340,4 @@ fn parse_json_array(
         .into_iter()
         .map(|obj| TantivyDocument::from_json_object(schema, obj))
         .collect::<Result<Vec<_>, _>>()?)
-}
-
-/// Macro to downcast an `ArrayRef` to concrete Arrow array type or return Err.
-///
-/// Users should check type-compatibility beforehand using [`ArrayRef::data_type`].
-macro_rules! downcast_array {
-    ($ARRAY:expr, $TY:ty) => {
-        $ARRAY.as_any().downcast_ref::<$TY>().ok_or_else(|| {
-            ArrowError::CastError(format!("Expected arrow array of type {}", stringify!($TY)))
-        })?
-    };
-}
-
-#[allow(clippy::too_many_lines)]
-pub fn array_to_terms(field: Field, arr: &ArrayRef) -> Result<Vec<Term>, ArrowError> {
-    let mut terms = Vec::with_capacity(arr.len());
-
-    match arr.data_type() {
-        // --- Floats → f64
-        DataType::Float16 => {
-            let a = downcast_array!(arr, Float16Array);
-            for i in 0..a.len() {
-                if a.is_valid(i) {
-                    let v = f64::from(a.value(i).to_f32());
-                    terms.push(Term::from_field_f64(field, v));
-                }
-            }
-        }
-        DataType::Float32 => {
-            let a = downcast_array!(arr, Float32Array);
-            for i in 0..a.len() {
-                if a.is_valid(i) {
-                    let v = f64::from(a.value(i));
-                    terms.push(Term::from_field_f64(field, v));
-                }
-            }
-        }
-        DataType::Float64 => {
-            let a = downcast_array!(arr, Float64Array);
-            for i in 0..a.len() {
-                if a.is_valid(i) {
-                    terms.push(Term::from_field_f64(field, a.value(i)));
-                }
-            }
-        }
-
-        // --- Unsigned ints → u64
-        DataType::UInt8 => {
-            let a = downcast_array!(arr, UInt8Array);
-            for i in 0..a.len() {
-                if a.is_valid(i) {
-                    terms.push(Term::from_field_u64(field, u64::from(a.value(i))));
-                }
-            }
-        }
-        DataType::UInt16 => {
-            let a = downcast_array!(arr, UInt16Array);
-            for i in 0..a.len() {
-                if a.is_valid(i) {
-                    terms.push(Term::from_field_u64(field, u64::from(a.value(i))));
-                }
-            }
-        }
-        DataType::UInt32 => {
-            let a = downcast_array!(arr, UInt32Array);
-            for i in 0..a.len() {
-                if a.is_valid(i) {
-                    terms.push(Term::from_field_u64(field, u64::from(a.value(i))));
-                }
-            }
-        }
-        DataType::UInt64 => {
-            let a = downcast_array!(arr, UInt64Array);
-            for i in 0..a.len() {
-                if a.is_valid(i) {
-                    terms.push(Term::from_field_u64(field, a.value(i)));
-                }
-            }
-        }
-
-        // --- Signed ints → i64
-        DataType::Int8 => {
-            let a = downcast_array!(arr, Int8Array);
-            for i in 0..a.len() {
-                if a.is_valid(i) {
-                    terms.push(Term::from_field_i64(field, i64::from(a.value(i))));
-                }
-            }
-        }
-        DataType::Int16 => {
-            let a = downcast_array!(arr, Int16Array);
-            for i in 0..a.len() {
-                if a.is_valid(i) {
-                    terms.push(Term::from_field_i64(field, i64::from(a.value(i))));
-                }
-            }
-        }
-        DataType::Int32 => {
-            let a = downcast_array!(arr, Int32Array);
-            for i in 0..a.len() {
-                if a.is_valid(i) {
-                    terms.push(Term::from_field_i64(field, i64::from(a.value(i))));
-                }
-            }
-        }
-        DataType::Int64 => {
-            let a = downcast_array!(arr, Int64Array);
-            for i in 0..a.len() {
-                if a.is_valid(i) {
-                    terms.push(Term::from_field_i64(field, a.value(i)));
-                }
-            }
-        }
-
-        // --- Boolean
-        DataType::Boolean => {
-            let a = downcast_array!(arr, BooleanArray);
-            for i in 0..a.len() {
-                if a.is_valid(i) {
-                    terms.push(Term::from_field_bool(field, a.value(i)));
-                }
-            }
-        }
-
-        // --- Dates
-        DataType::Date32 => {
-            let a = downcast_array!(arr, Date32Array);
-            for i in 0..a.len() {
-                if a.is_valid(i) {
-                    terms.push(Term::from_field_date(
-                        field,
-                        tantivy::DateTime::from_timestamp_secs(i64::from(a.value(i)) * 86_400),
-                    ));
-                }
-            }
-        }
-        DataType::Date64 => {
-            let a = downcast_array!(arr, Date64Array);
-            for i in 0..a.len() {
-                if a.is_valid(i) {
-                    terms.push(Term::from_field_date(
-                        field,
-                        tantivy::DateTime::from_timestamp_millis(a.value(i)),
-                    ));
-                }
-            }
-        }
-
-        // --- UTF8 text
-        DataType::Utf8 => {
-            let a = downcast_array!(arr, StringArray);
-            for i in 0..a.len() {
-                if a.is_valid(i) {
-                    terms.push(Term::from_field_text(field, a.value(i)));
-                }
-            }
-        }
-        DataType::LargeUtf8 => {
-            let a = downcast_array!(arr, LargeStringArray);
-            for i in 0..a.len() {
-                if a.is_valid(i) {
-                    terms.push(Term::from_field_text(field, a.value(i)));
-                }
-            }
-        }
-
-        // --- Binary blobs
-        DataType::Binary => {
-            let a = downcast_array!(arr, BinaryArray);
-            for i in 0..a.len() {
-                if a.is_valid(i) {
-                    terms.push(Term::from_field_bytes(field, a.value(i)));
-                }
-            }
-        }
-        DataType::LargeBinary => {
-            let a = downcast_array!(arr, LargeBinaryArray);
-            for i in 0..a.len() {
-                if a.is_valid(i) {
-                    terms.push(Term::from_field_bytes(field, a.value(i)));
-                }
-            }
-        }
-
-        // --- Everything else is unsupported
-        other => {
-            return Err(ArrowError::NotYetImplemented(format!(
-                "Cannot use primary key of arrow type {other:?} for full-text search"
-            )));
-        }
-    }
-
-    Ok(terms)
 }
